Rename p1000e2e package
